+++ src/extendedBitStream.ts
@@ -1,0 +1,86 @@
+import { BitStream } from "bit-buffer";
+
+export default class ExtendedBitStream extends BitStream {
+
+    constructor(buffer: Buffer) {
+        super(buffer, buffer.byteOffset, buffer.byteLength);
+    }
+
+    public swapEndian() {
+        this.bigEndian = !this.bigEndian;
+    }
+
+    // the type definition for BitStream does not include the _index property
+    // since it's supposed to be private, but it's needed 4 times here sooo
+
+    public alignByte() {
+        // @ts-expect-error
+        const nextClosestByteIndex = 8 * Math.ceil(this._index / 8)
+        // @ts-expect-error
+        const bitDistance = nextClosestByteIndex - this._index;
+
+        this.skipBits(bitDistance);
+    }
+
+    public bitSeek(bitPos: number) {
+        // @ts-expect-error
+        this._index = bitPos;
+    }
+
+    public skipBits(bitCount: number) {
+        // @ts-expect-error
+        this._index += bitCount;
+    }
+
+    public skipBytes(bytes: number) {
+        const bits = bytes * 8;
+        this.skipBits(bits);
+    }
+
+    public skipBit() {
+        this.skipBits(1);
+    }
+
+    public skipInt8() {
+        this.skipBytes(1);
+    }
+
+    public skipInt16() {
+        // Skipping a uint16 is the same as skipping 2 uint8's
+        this.skipBytes(2);
+    }
+
+    public readBit() {
+        return this.readBits(1);
+    }
+
+    public readBytes(length: number) {
+        return Array(length).fill(0).map(() => this.readUint8());
+    }
+
+    public readBuffer(length: number) {
+        return Buffer.from(this.readBytes(length));
+    }
+
+    public readUTF16String(length: number) {
+        return this.readBuffer(length).toString('utf16le').replace(/\0.*$/, '');
+    }
+
+    public writeBit(bit: number) {
+        this.writeBits(bit, 1);
+    }
+
+
+    public writeBuffer(buffer: Buffer) {
+        buffer.forEach(byte => this.writeUint8(byte));
+    }
+
+    public writeUTF16String(string: string) {
+        const stringBuffer = Buffer.from(string, 'utf16le');
+        const terminatedBuffer = Buffer.alloc(0x14);
+
+        stringBuffer.copy(terminatedBuffer);
+
+        this.writeBuffer(terminatedBuffer);
+    }
+}
